# Parallel streams can be 8 times faster. 