Using:
java version "1.8.0_102"
Java(TM) SE Runtime Environment (build 1.8.0_102-b14)
Java HotSpot(TM) 64-Bit Server VM (build 25.102-b14, mixed mode)

Tests run on an 2015 15' maxbook pro ( quad core I7 at 2.5Ghz)

And:
- used Math.max instead of Integer.max
- added extra cases for stream.max() for timing
- added collections foreach case
- added int[] and IntStream.of() for comparisons.

## Conclusion
The average time for the parrallel stream on an IntStream.of([]) is approximately 8 times faster than the traditional for loop.
- Using Math.max is not necessarily faster.
- increasing the steam size causes better parrallel stream performances.
- using IntStream.of is a lot faster faster.
- using Stream.max is faster than aggregate.
- using the foreach on ArrayList collection is seriously slow.